Extract isModuleResizable into a separate file
