+++ app/src/editor/canvas/components/Module.jsx
@@ -17,7 +17,8 @@ import * as RunController from './RunController'
 
 import ModuleStyles from '$editor/shared/components/Module.pcss'
 import styles from './Module.pcss'
-import { Resizer, isModuleResizable } from './Resizer'
+import { Resizer } from './Resizer'
+import isModuleResizable from '$editor/canvas/utils/isModuleResizable'
 
 class CanvasModule extends React.PureComponent {
     static contextType = RunController.Context
